Simplifications and sign convention change
I also changed the sign convention of torsion to be positive counter-clockwise on the right side of a section cut. This is consistent with the the right hand rule.

I simplified the torsional functions since torsion is constant on each beam segment. This will eventually have to change when distributed torsional moments are supported.